refactor from code review
